data/requirements.txt
numpy
pandas
tabula-py
jpype1
